#5 refactor: enums package로 분리
