+++ src/main/java/com/world/playstay/user/enums/AuthStatus.java
@@ -1,0 +1,6 @@
+package com.world.playstay.user.enums;
+
+public enum AuthStatus {
+  UNAUTHENTICATED,
+  AUTHENTICATED,
+}
