+++ src/main/java/com/world/playstay/user/enums/MemberShipStatus.java
@@ -1,0 +1,7 @@
+package com.world.playstay.user.enums;
+
+public enum MemberShipStatus {
+  BASIC,
+  STANDARD,
+  PREMIUM
+}
