+++ src/main/java/com/world/playstay/user/enums/UserType.java
@@ -1,0 +1,6 @@
+package com.world.playstay.user.enums;
+
+public enum UserType {
+  HOST,
+  GUEST,
+}
