+++ src/main/java/com/world/playstay/user/service/GuestService.java
@@ -2,8 +2,8 @@ package com.world.playstay.user.service;
 
 import com.world.playstay.global.util.HashUtil;
 import com.world.playstay.user.dao.GuestMapper;
-import com.world.playstay.user.entity.AuthStatus;
 import com.world.playstay.user.entity.Guest;
+import com.world.playstay.user.enums.AuthStatus;
 import com.world.playstay.user.exception.DuplicatedUserException;
 import com.world.playstay.user.exception.UserNotFoundException;
 import java.util.List;
