+++ src/test/java/com/world/playstay/user/service/HostServiceTest.java
@@ -9,9 +9,9 @@ import static org.mockito.Mockito.when;
 
 import com.world.playstay.ServiceTest;
 import com.world.playstay.user.dao.HostMapper;
-import com.world.playstay.user.entity.AuthStatus;
 import com.world.playstay.user.entity.Host;
-import com.world.playstay.user.entity.Host.MemberShipStatus;
+import com.world.playstay.user.enums.AuthStatus;
+import com.world.playstay.user.enums.MemberShipStatus;
 import com.world.playstay.user.exception.DuplicatedUserException;
 import com.world.playstay.user.exception.UserNotFoundException;
 import java.util.Optional;
